General cleanup, no semantic changes
