+++ src/main/scala/com/foursquare/rogue/RogueException.scala
